Seperated Demos from KinectManager
