+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Prefabs/Demo.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: c0c5115083c9706408b4c89ed3c25849
+folderAsset: yes
+timeCreated: 1511207321
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
