+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Prefabs/Demo/TrackingCapsule.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 3c1da26f09d22214d8d2c7deca3c53be
+timeCreated: 1511207391
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
