+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Prefabs/Demo/TrackingCapsule.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: c85cb56143f6cd04fb47f302c22179da
+timeCreated: 1511207340
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
