+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Scripts/Demo.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 69a25c3636a36cf43a900acbc01c47f0
+folderAsset: yes
+timeCreated: 1511208829
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
