+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Scripts/Demo/BlobTrackerDemo.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 7577e1fac30c919449d15e56bf4156dc
+timeCreated: 1511208053
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
