+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Scripts/Demo/FaceTrackerOpenCVDemo.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: c730c3a3136a755478c1b90d5c0ac132
+timeCreated: 1511208997
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
