+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Scripts/KinectCVUtilities.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e229e6549b06f094680bac8d03b85434
+timeCreated: 1511208150
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
