+++ PlatformDemos/NetworkItUnity+KinectV2/NetworkItUnity+KinectV2/Assets/Scripts/Views/IRView.cs
@@ -17,6 +17,6 @@ public class IRView : MonoBehaviour
             return;
         }
 
-        gameObject.GetComponent<Renderer>().material.mainTexture = infraredSource.GetInfraredTexture();
+        gameObject.GetComponent<Renderer>().material.mainTexture = infraredSource.GetIRTexture();
     }
 }
